+++ src/cobib/commands/modify.py
@@ -351,7 +351,7 @@ class ModifyCommand(Command):
                             path = RelPath(file)
                             if path.path.stem == label:
                                 LOGGER.info("Also renaming associated file '%s'.", str(path))
-                                target = RelPath(path.path.parent / f"{entry.label}.pdf")
+                                target = RelPath(path.path.with_stem(entry.label))
                                 if target.path.exists():
                                     LOGGER.warning(  # pragma: no cover
                                         "Found conflicting file, not renaming '%s'.", str(path)